Depot
Timestamp fields in float, not int
FileObject magic functions

+++ hworker/depot/database/models.py
@@ -72,9 +72,9 @@ class CheckResult(Base):
     rating: Mapped[float] = mapped_column(Float)
     category: Mapped[CheckCategoryEnum] = mapped_column(Enum(CheckCategoryEnum))
     check_ID: Mapped[str] = mapped_column(String)
-    check_timestamp: Mapped[int] = mapped_column(Integer)
+    check_timestamp: Mapped[float] = mapped_column(Float)
     solution_ID: Mapped[str] = mapped_column(String)
-    solution_timestamp: Mapped[int] = mapped_column(Integer)
+    solution_timestamp: Mapped[float] = mapped_column(Float)
     verdict: Mapped[VerdictEnum] = mapped_column(Enum(VerdictEnum))
     stdout: Mapped[bytes] = mapped_column(LargeBinary)
     stderr: Mapped[bytes] = mapped_column(LargeBinary)
@@ -85,9 +85,9 @@ class CheckResult(Base):
         rating: float = None,
         category: CheckCategoryEnum = None,
         check_ID: str = None,
-        check_timestamp: int = None,
+        check_timestamp: float = None,
         solution_ID: str = None,
-        solution_timestamp: int = None,
+        solution_timestamp: float = None,
         verdict: VerdictEnum = None,
         stdout: bytes = None,
         stderr: bytes = None,
